Move PreProcessing to new package
This brings it in line with the post processing
